`CredentialDatabase`
