+++ build.gradle
@@ -29,21 +29,6 @@ subprojects {
     dependencies {
         compileOnly "org.jetbrains:annotations:26.0.2"
         annotationProcessor(compileOnly("org.projectlombok:lombok:1.18.36"))
-
-        testImplementation platform('org.junit:junit-bom:5.11.4')
-        testImplementation 'org.junit.jupiter:junit-jupiter'
-        testAnnotationProcessor(testCompileOnly("org.projectlombok:lombok:1.18.36"))
-    }
-
-    test {
-        useJUnitPlatform()
-
-        ignoreFailures true
-
-        testLogging {
-            exceptionFormat = "full"
-            events "passed", "skipped", "failed"
-        }
     }
 
     idea {
@@ -62,10 +47,6 @@ subprojects {
 
     tasks.withType(JavaCompile).configureEach {
         options.encoding = "UTF-8"
-    }
-
-    tasks.withType(Test).configureEach { tsk ->
-        tsk.systemProperty "file.encoding", "UTF-8"
     }
 
     compileJava.dependsOn clean
