+++ mysql/src/main/java/io/github/blockneko11/simpledbc/impl/MySQLDatabase.java
@@ -6,37 +6,12 @@ import org.jetbrains.annotations.NotNull;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public class MySQLDatabase extends AbstractDatabase {
-    private final String username;
-    private final String password;
-    private final String databaseName;
-
+public class MySQLDatabase extends CredentialDatabase {
     public MySQLDatabase(@NotNull String url,
                          @NotNull String username,
                          @NotNull String password,
                          @NotNull String databaseName) {
-        super(url);
-        this.username = username;
-        this.password = password;
-        this.databaseName = databaseName;
-    }
-
-    @NotNull
-    @Override
-    public String getUsername() {
-        return this.username;
-    }
-
-    @NotNull
-    @Override
-    public String getPassword() {
-        return this.password;
-    }
-
-    @NotNull
-    @Override
-    public String getDatabaseName() {
-        return this.databaseName;
+        super(url, username, password, databaseName);
     }
 
     @Override
@@ -49,9 +24,9 @@ public class MySQLDatabase extends AbstractDatabase {
         if (!isConnected()) {
             String uri = StringUtil.format("jdbc:mysql://{0}/{1}?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC",
                     url,
-                    this.databaseName
+                    databaseName
             );
-            setConnection(DriverManager.getConnection(uri, this.username, this.password));
+            setConnection(DriverManager.getConnection(uri, username, password));
         }
     }
 }
